refactor: shell_ui.rsからURL intercept処理をapp_frame.rsへ抽出 (ロジック行数をさらに削減)
